refactor:
alters the multiple client scenario to use a metadata flag to switch behaviour, rather than
inheriting
