+++ features/multi_client.feature
@@ -4,10 +4,11 @@ Feature: Multi-client support
     configuration options. A report which is captured by a given client should
     use the correct API key and configuration options.
 
-    Scenario: An unhandled error captured while offline is detected by two clients with different API keys
+Scenario: An unhandled error captured while offline is detected by two clients with different API keys
     When I run "MultiClientApiKeyScenario" with the defaults
     When I force stop the "com.bugsnag.android.mazerunner" Android app
-    And I set environment variable "EVENT_TYPE" to "MultiClientScenario"
+    And I set environment variable "EVENT_TYPE" to "MultiClientApiKeyScenario"
+    And I set environment variable "EVENT_METADATA" to "DeliverReport"
     And I start the "com.bugsnag.android.mazerunner" Android app using the "com.bugsnag.android.mazerunner.MainActivity" activity
     Then I should receive 2 requests
